+++ servers/sunrpc/handshake_prog.go
@@ -1,4 +1,4 @@
-package sunrpcserver
+package sunrpc
 
 import (
 	"fmt"
